requirements.txt
alembic==1.11.1
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.15.0
Flask-Migrate==2.7.0
flask-restplus==0.13.0
flask-restx==1.1.0
Flask-Script==2.0.5
Flask-SQLAlchemy==3.0.5
GeoAlchemy2==0.13.3
greenlet==2.0.2
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
packaging==23.1
psycopg2-binary==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Werkzeug==2.3.6
zipp==3.15.0

pytest~=7.4.0